Rename calc files
